backend/pyproject.toml
[tool.poetry]
name = "pokemon-battle-api"
version = "1.0.0"
description = "A turn-based Pokemon battle simulation API"
authors = ["Shayesteh HS"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = ">=5.0"
djangorestframework = ">=3.14"
django-filter = ">=23.5"
uuid7 = ">=0.1.0"
djangorestframework-simplejwt = ">=5.3"
drf-spectacular = ">=0.27"
drf-to-mkdoc = ">=0.2"
psycopg = {extras = ["binary"], version = ">=3.1"}
dj-database-url = ">=2.1"
redis = ">=5.0"
httpx = ">=0.27"
gunicorn = ">=21.0"
mkdocs = ">=1.5"
mkdocs-material = ">=9.5"
coreapi = ">=2.3.3"
django-cors-headers = ">=4.3"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0"
pytest-django = ">=4.8"
pytest-cov = ">=4.1"
ruff = ">=0.3"
pre-commit = ">=3.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "DJ",     # flake8-django
    "TID",    # flake8-tidy-imports
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.lint.isort]
known-first-party = ["config", "players", "pokemon", "battles", "scoreboard", "logs", "utils"]

[tool.coverage.run]
source = ["."]
omit = ["*/migrations/*", "*/tests/*", "manage.py", "conftest.py"]
